+++ CHANGELOG.md
@@ -2,3 +2,4 @@ Update 1.01
 -----------
 * Fixed a little bug in ProgressArcDrawable bounds which caused the arc to get drawn behind the fab in some cases.
 * Added `app:circleSize` custom attribute with values `normal` or `mini` to fix some layout weird measures.
+* Added custom attribute `app:roundedStroke="true"` to allow a rounded stroke on the progress circle heads.